services/workers/pyannote_audio_service/pyproject.toml
[project]
name = "pyannote-audio-service"
description = "State-of-the-art speaker diarization toolkit"
readme = "README.md"
authors = [{ name = "Hervé BREDIN", email = "herve@pyannote.ai" }]
requires-python = ">=3.10"

dynamic = ["version"]

dependencies = [
    "asteroid-filterbanks>=0.4.0",
    "einops>=0.8.1",
    "huggingface-hub>=0.28.1",
    "lightning>=2.4",
    "opentelemetry-api>=1.34.0",
    "opentelemetry-sdk>=1.34.0",
    "opentelemetry-exporter-otlp>=1.34.0",
    "pyannote-core>=6.0.1",
    "pyannote-database>=6.0.0",
    "pyannote-metrics>=4.0.0",
    "pyannote-pipeline>=4.0.0",
    "pytorch-metric-learning>=2.8.1",
    "rich>=13.9.4",
    "safetensors>=0.5.2",
    "torch-audiomentations>=0.12.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchcodec>=0.6.0",
    "torchmetrics>=1.6.1",
    "soundfile>=0.13.1",
    "matplotlib>=3.10.0",
    "pyannoteai-sdk>=0.3.0",
]

[project.scripts]
pyannote-audio = "pyannote.audio.__main__:app"

[project.optional-dependencies]
test = ["papermill>=2.6.0", "pytest>=8.3.4", "speechbrain>=1.0.2"]
doc = ["sphinx-rtd-theme>=3.0.2", "sphinx>=8.1.3"]
cli = ["typer>=0.15.1"]
separation = ["asteroid>=0.5.2", "transformers>=4.48.3"]
dev = ["ipython>=8.32.0"]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/pyannote"]

[dependency-groups]
dev = ["ipykernel>=6.29.5"]
